all: retrofit under io.orijtech.integrations.ocjdbc
Retrofit this package under
    "io.orijtech.integrations.ocjdbc"
instead of
    "io.opencensus.integration.jdbc"

because publishing changes under the official OpenCensus
artifact group is a source of contention due to the OpenCensus Java
team not wanting to have to maintain integrations for life but also
the process of publishing is highly bureaucratic.
Also we've waited for more than 3 months for a single release which
has stalled usage, promotion, feedback and marketing of this integration
as per:
* https://github.com/census-ecosystem/opencensus-java-jdbc/issues/32
* https://github.com/census-ecosystem/opencensus-java-jdbc/issues/56
* https://github.com/census-instrumentation/opencensus-website/issues/527

This package to be published is maintained under good faith for the entire OpenCensus
community and the licenses all stand; just that the artifact is now
being published under "io.orijtech.integrations.ocjdbc" which guts out
on the previously mentioned bureaucracy, indefinite stalls in publishing
and puts this integration into the hands of users.